% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{calc_sample_size_kappa}
\alias{calc_sample_size_kappa}
\title{Calculate Sample Size for Kappa Statistic}
\usage{
calc_sample_size_kappa(
  kappa0 = 0.4,
  kappa1 = 0.6,
  props = c(0.5, 0.5),
  alpha = 0.05,
  power = 0.8,
  raters = 2
)
}
\arguments{
\item{kappa0}{Null hypothesis value of kappa (H0).}

\item{kappa1}{Alternative hypothesis value of kappa (H1).}

\item{props}{Expected proportions of categories. Must sum to 1.}

\item{alpha}{Significance level. Default 0.05.}

\item{power}{Desired power. Default 0.8.}

\item{raters}{Number of raters (>=2). Default 2.}
}
\value{
A list with sample size n and parameters.
}
\description{
Calculate required sample size for consistency tests.
Directly wraps kappaSize::PowerBinary / Power3Cats / Power4Cats / Power5Cats.
}
\examples{
calc_sample_size_kappa(kappa0 = 0.4, kappa1 = 0.6, props = c(0.5, 0.5))
calc_sample_size_kappa(kappa0 = 0.4, kappa1 = 0.6,
                       props = c(0.6, 0.3, 0.1), raters = 3)
}
