% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_kappa_data}
\alias{check_kappa_data}
\title{Check Kappa Data Validity}
\usage{
check_kappa_data(data, type = "cohen")
}
\arguments{
\item{data}{A data frame or matrix.}

\item{type}{Type of kappa analysis.}
}
\value{
Logical indicating validity.
}
\description{
Internal function to validate input data for kappa analysis.
}
\keyword{internal}
