% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpret_kappa}
\alias{interpret_kappa}
\title{Interpret Kappa Value}
\usage{
interpret_kappa(kappa)
}
\arguments{
\item{kappa}{Numeric value of Kappa statistic.}
}
\value{
A named list with level, description, and color code.
}
\description{
Interpret the strength of agreement based on Landis and Koch criteria.
}
\examples{
interpret_kappa(0.3)
interpret_kappa(0.75)
}
