% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{print.cate_analysis}
\alias{print.cate_analysis}
\title{Print Method for cate_analysis Objects}
\usage{
\method{print}{cate_analysis}(x, ...)
}
\arguments{
\item{x}{An object of class cate_analysis.}

\item{...}{Additional arguments.}
}
\description{
Print Method for cate_analysis Objects
}
