% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{print.cate_design}
\alias{print.cate_design}
\title{Print Method for cate_design Objects}
\usage{
\method{print}{cate_design}(x, ...)
}
\arguments{
\item{x}{An object of class cate_design.}

\item{...}{Additional arguments.}
}
\description{
Print Method for cate_design Objects
}
