% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{print.cate_fixed_n}
\alias{print.cate_fixed_n}
\title{Print Method for cate_fixed_n Objects}
\usage{
\method{print}{cate_fixed_n}(x, ...)
}
\arguments{
\item{x}{An object of class cate_fixed_n.}

\item{...}{Additional arguments.}
}
\description{
Print Method for cate_fixed_n Objects
}
