% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{run_cate_app}
\alias{run_cate_app}
\title{Run CATE Shiny Application}
\usage{
run_cate_app(
  port = getOption("shiny.port"),
  launch.browser = getOption("shiny.launch.browser", interactive()),
  host = getOption("shiny.host", "127.0.0.1")
)
}
\arguments{
\item{port}{The TCP port for the application. Defaults to random available port.}

\item{launch.browser}{Logical. Whether to launch browser automatically.}

\item{host}{The IPv4 address to listen on.}
}
\value{
A Shiny application object (invisible).
}
\description{
Launch the Shiny application for design and analysis of consistency tests
based on Kappa statistic with categorical responses.
}
\examples{
if(interactive()){
run_cate_app()
}
}
