% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summary.cate_analysis}
\alias{summary.cate_analysis}
\title{Summary Method for cate_analysis Objects}
\usage{
\method{summary}{cate_analysis}(object, ...)
}
\arguments{
\item{object}{An object of class cate_analysis.}

\item{...}{Additional arguments.}
}
\description{
Summary Method for cate_analysis Objects
}
